EIP package to Standards package
