--- eip/doc.go
@@ -1,2 +1,0 @@
-// Package eip provides structures and functions to represent and manipulate Ethereum Improvement Proposals (EIPs) and Ethereum standards.
-package eip
