+++ standards/doc.go
@@ -1,0 +1,2 @@
+// Package standards provides structures and functions to represent and manipulate Ethereum Improvement Proposals (EIPs) and Ethereum standards.
+package standards
