+++ standards/errors.go
@@ -1,0 +1,9 @@
+package standards
+
+import "errors"
+
+var (
+
+	// ErrStandardNotFound is returned when a standard is not found.
+	ErrStandardNotFound = errors.New("standard not found")
+)
