+++ standards/storage.go
@@ -1,4 +1,4 @@
-package eip
+package standards
 
 import (
 	"fmt"
